Fixed selector for ECC transition buttons.
The JQuery selector was too vague, and only looked for the
source-ctrl-btn class. The config choice button also has this class, so
it caused a conflict where the config choice button had a click handler
in addition to its href.

I fixed this by adding ids to each button and selecting on those.